Fixed VS compilation error.
git-svn-id: svn+ssh://svn.code.sf.net/p/supertuxkart/code/main/branches/hilnius@13569 178a84e3-b1eb-0310-8ba1-8eac791a3b58

+++ src/network/stk_host.cpp
@@ -190,7 +190,7 @@ uint8_t* STKHost::receiveRawPacket(TransportAddress* sender)
     }
     struct sockaddr_in *sin = (struct sockaddr_in *) (&addr);
     // we received the data
-    sender->ip = turnEndianness(sin->sin_addr.s_addr);
+    sender->ip = turnEndianness((uint32_t)(sin->sin_addr.s_addr));
     sender->port = turnEndianness(sin->sin_port);
 
     if (addr.sa_family == AF_INET)
